x/processor/keeper/query_show_processor.go
package keeper

import (
	"context"
	"intellix/x/processor/types"

	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) ShowProcessor(goCtx context.Context, req *types.QueryShowProcessorRequest) (*types.QueryShowProcessorResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}

	ctx := sdk.UnwrapSDKContext(goCtx)
	processor, found := k.GetProcessor(ctx, req.Id)
	if !found {
		return nil, sdkerrors.ErrKeyNotFound
	}

	return &types.QueryShowProcessorResponse{Processor: processor}, nil
}
